tessellation and geometry displacement

+++ engine/editor/CustomControls.h
@@ -6,7 +6,7 @@ namespace ImGui
     bool GColorEdit3(const char* label, glm::vec3& value);
     bool GColorEdit4(const char* label, glm::vec4& value);
     bool GDragFloat(const char* label, float* value, float step = 0.01f, float min = 0.0f, float max = 0.0f);
-    bool GDragInt(const char* label, int* value, int step = 1, int min = 0, int max = 0);
+    bool GDragInt(const char* label, int* value, float step = 1.f, int min = 0, int max = 0);
     bool GDragFloatVec3(const char* label, glm::vec3& values, float step = 0.01f, float min = 0.0f, float max = 0.0f);
     bool GDragTransform(glm::vec3& pos, glm::vec3& rot, glm::vec3& scale, float step = 0.01f, float min = 0.0f, float max = 0.0f);
     bool GAssetHolder(const char* label, const std::string& source);
